Modularize services architecture
- Categorized services into browser/, ai/, features/ subdirs
- Moved files: browser (action, application), ai (ollama, model-settings, etc.), features (kb, ocr, etc.)
- Added barrel exports in services/index.ts for clean imports
- Updated internal imports in moved files
- Improved separation of concerns for maintainability
