+++ memory-bank/arch-improvements-enhance-routes-ui.md
@@ -71,4 +71,12 @@ Group routes/settings; implement ChatInputShell wrappers; enforce styling conven
 
 ## Status
 
-Pending
+Completed
+
+## Post-Completion Notes
+
+- Settings routes grouped in routes/settings/
+- Shared SettingsLayout.tsx with tabs and lazy loading for sections
+- ChatInputShell.tsx and ChatTextarea.tsx created for reduced duplication
+- Data-testid added for automation
+- Nested routes implemented for settings
