+++ src/services/ai/chrome.ts
