+++ src/services/ai/model-settings.ts
