+++ src/services/ai/ollama.ts
@@ -1,12 +1,12 @@
 import { Storage } from "@plasmohq/storage"
-import { cleanUrl } from "../libs/clean-url"
-import { urlRewriteRuntime } from "../libs/runtime"
+import { cleanUrl } from "@/libs/clean-url"
+import { urlRewriteRuntime } from "@/libs/runtime"
 import { getChromeAIModel } from "./chrome"
 import {
   getOllamaEnabled,
   setNoOfRetrievedDocs,
   setTotalFilePerKB
-} from "./app"
+} from "../features/app"
 import fetcher from "@/libs/fetcher"
 import { ollamaFormatAllCustomModels } from "@/db/dexie/models"
 import { getAllModelNicknames } from "@/db/dexie/nickname"
