+++ src/services/index.ts
@@ -1,0 +1,18 @@
+// Browser services
+export * from "./browser/action"
+export * from "./browser/application"
+
+// AI services
+export * from "./ai/chrome"
+export * from "./ai/elevenlabs"
+export * from "./ai/model-settings"
+export * from "./ai/ollama"
+export * from "./ai/openai-tts"
+export * from "./ai/tts"
+
+// Feature services
+export * from "./features/app"
+export * from "./features/kb"
+export * from "./features/ocr"
+export * from "./features/search"
+export * from "./features/title"
